+++ src/app/_constants/landing.ts
