+++ src/app/_types/landing.type.ts
